1. decouple evaluate and train
