+++ examples/few_shot/p-tuning/ptuning.py
@@ -29,8 +29,10 @@ from paddlenlp.data import Stack, Tuple, Pad
 from paddlenlp.datasets import load_dataset
 from paddlenlp.transformers import LinearDecayWithWarmup
 
-from data import read_fn_dict, convert_example, create_dataloader
 from model import ErnieForPretraining, ErnieMLMCriterion
+from data import read_fn_dict, create_dataloader
+from data import convert_example, convert_chid_example
+from evaluate import do_evaluate, do_evaluate_chid
 
 # yapf: disable
 parser = argparse.ArgumentParser()
@@ -64,65 +66,6 @@ def set_seed(seed):
     paddle.seed(seed)
 
 
-@paddle.no_grad()
-def do_evaluate(model, tokenizer, data_loader, label_normalize_dict):
-    model.eval()
-
-    total_num = 0
-    correct_num = 0
-
-    normed_labels = [
-        normalized_lable
-        for origin_lable, normalized_lable in label_normalize_dict.items()
-    ]
-
-    label_length = len(normed_labels[0])
-
-    for batch in data_loader:
-        src_ids, token_type_ids, masked_positions, masked_lm_labels = batch
-
-        # [bs * label_length, vocab_size]
-        prediction_probs = model.predict(
-            input_ids=src_ids,
-            token_type_ids=token_type_ids,
-            masked_positions=masked_positions)
-
-        batch_size = len(src_ids)
-        vocab_size = prediction_probs.shape[1]
-
-        #prediction_probs: [batch_size, label_lenght, vocab_size]
-        prediction_probs = paddle.reshape(
-            prediction_probs, shape=[batch_size, -1, vocab_size]).numpy()
-
-        # [label_num, label_length]
-        label_ids = np.array(
-            [tokenizer(label)["input_ids"][1:-1] for label in normed_labels])
-
-        y_pred = np.ones(shape=[batch_size, len(label_ids)])
-
-        # calculate joint distribution of candidate labels
-        for index in range(label_length):
-            y_pred *= prediction_probs[:, index, label_ids[:, index]]
-
-        # get max probs label's index
-        y_pred_index = np.argmax(y_pred, axis=-1)
-
-        y_true_index = []
-        for masked_lm_label in masked_lm_labels.numpy():
-            label_text = "".join(
-                tokenizer.convert_ids_to_tokens(list(masked_lm_label)))
-
-            label_index = normed_labels.index(label_text)
-            y_true_index.append(label_index)
-
-        y_true_index = np.array(y_true_index)
-
-        total_num += len(y_true_index)
-        correct_num += (y_true_index == y_pred_index).sum()
-
-    return 100 * correct_num / total_num, total_num
-
-
 def do_train():
     paddle.set_device(args.device)
     rank = paddle.distributed.get_rank()
@@ -147,6 +90,10 @@ def do_train():
         label_norm_dict = json.load(f)
 
     read_fn = read_fn_dict[args.task_name]
+
+    convert_example_fn = convert_example if args.task_name != "chid" else convert_chid_example
+    evaluate_fn = do_evaluate if args.task_name != "chid" else do_evaluate_chid
+
     train_ds = load_dataset(
         read_fn,
         data_path=train_set,
@@ -169,19 +116,29 @@ def do_train():
 
     tokenizer = ppnlp.transformers.ErnieTokenizer.from_pretrained('ernie-1.0')
 
+    if args.task_name != "chid":
+        # [src_ids, token_type_ids, masked_positions, masked_lm_labels]
+        batchify_fn = lambda samples, fn=Tuple(
+            Pad(axis=0, pad_val=tokenizer.pad_token_id),  # src_ids
+            Pad(axis=0, pad_val=tokenizer.pad_token_type_id),  # token_type_ids
+            Stack(dtype="int64"),  # masked_positions
+            Stack(dtype="int64"),  # masked_lm_labels
+        ): [data for data in fn(samples)]
+    else:
+        # [src_ids, token_type_ids, masked_positions, masked_lm_labels, candidate_labels_ids]
+        batchify_fn = lambda samples, fn=Tuple(
+            Pad(axis=0, pad_val=tokenizer.pad_token_id),  # src_ids
+            Pad(axis=0, pad_val=tokenizer.pad_token_type_id),  # token_type_ids
+            Stack(dtype="int64"),  # masked_positions
+            Stack(dtype="int64"),  # masked_lm_labels
+            Stack(dtype="int64"),  # candidate_labels_ids [candidate_num, label_length]
+        ): [data for data in fn(samples)]
+
     trans_func = partial(
-        convert_example,
+        convert_example_fn,
         tokenizer=tokenizer,
         max_seq_length=args.max_seq_length,
         p_embedding_num=args.p_embedding_num)
-
-    # [src_ids, token_type_ids, masked_positions, masked_lm_labels]
-    batchify_fn = lambda samples, fn=Tuple(
-        Pad(axis=0, pad_val=tokenizer.pad_token_id),  # src_ids
-        Pad(axis=0, pad_val=tokenizer.pad_token_type_id),  # token_type_ids
-        Stack(dtype="int64"),  # masked_positions
-        Stack(dtype="int64"),  # masked_lm_labels
-    ): [data for data in fn(samples)]
 
     train_data_loader = create_dataloader(
         train_ds,
@@ -258,11 +215,11 @@ def do_train():
             lr_scheduler.step()
             optimizer.clear_grad()
 
-        dev_accuracy, total_num = do_evaluate(model, tokenizer, dev_data_loader,
+        dev_accuracy, total_num = evaluate_fn(model, tokenizer, dev_data_loader,
                                               label_norm_dict)
         print("epoch:{}\tdev_accuracy:{:.3f}\ttotal_num:{}".format(
             epoch, dev_accuracy, total_num))
-        test_accuracy, total_num = do_evaluate(
+        test_accuracy, total_num = evaluate_fn(
             model, tokenizer, public_test_data_loader, label_norm_dict)
         print("epoch:{}\ttest_accuracy:{:.3f}\ttotal_num:{}".format(
             epoch, test_accuracy, total_num))
