+++ examples/few_shot/p-tuning/ptuning_chid.py
@@ -65,80 +65,6 @@ def set_seed(seed):
     paddle.seed(seed)
 
 
-@paddle.no_grad()
-def do_evaluate(model, tokenizer, data_loader, label_normalize_dict):
-    model.eval()
-
-    total_num = 0
-    correct_num = 0
-
-    normed_labels = [
-        normalized_lable
-        for origin_lable, normalized_lable in label_normalize_dict.items()
-    ]
-
-    label_length = len(normed_labels[0])
-
-    for batch in data_loader:
-        src_ids, token_type_ids, masked_positions, masked_lm_labels, candidate_label_ids = batch
-
-        # [bs * label_length, vocab_size]
-        prediction_probs = model.predict(
-            input_ids=src_ids,
-            token_type_ids=token_type_ids,
-            masked_positions=masked_positions)
-
-        batch_size = len(src_ids)
-        vocab_size = prediction_probs.shape[1]
-
-        #prediction_probs: [batch_size, label_lenght, vocab_size]
-        prediction_probs = paddle.reshape(
-            prediction_probs, shape=[batch_size, -1, vocab_size]).numpy()
-
-        candidate_num = candidate_label_ids.shape[1]
-
-        # [batch_size, candidate_num(7)]
-        y_pred = np.ones(shape=[batch_size, candidate_num])
-
-        for label_idx in range(candidate_num):
-
-            # [bathc_size, label_length(4)] 
-            single_candidate_label_ids = candidate_label_ids[:, label_idx, :]
-            # 计算每个候选 label 的联合概率分布
-            for index in range(label_length):
-                # [batch_size,]
-                slice_word_ids = single_candidate_label_ids[:, index].numpy()
-
-                batch_single_token_prob = []
-                for bs_index in range(batch_size):
-                    # [1, 1]
-                    single_token_prob = prediction_probs[
-                        bs_index, index, slice_word_ids[bs_index]]
-                    batch_single_token_prob.append(single_token_prob)
-
-                y_pred[:, label_idx] *= np.array(batch_single_token_prob)
-
-        # get max probs label's index
-        y_pred_index = np.argmax(y_pred, axis=-1)
-
-        y_true_index = []
-
-        for index, masked_lm_label in enumerate(masked_lm_labels.numpy()):
-            # [cantidate_num, label_length]
-            tmp_candidate_label_ids = candidate_label_ids[index, :, :]
-            for idx, label_ids in enumerate(tmp_candidate_label_ids.numpy()):
-                if np.equal(label_ids, masked_lm_label).all():
-                    y_true_index.append(idx)
-                    continue
-
-        y_true_index = np.array(y_true_index)
-
-        total_num += len(y_true_index)
-        correct_num += (y_true_index == y_pred_index).sum()
-
-    return 100 * correct_num / total_num, total_num
-
-
 def do_train():
     paddle.set_device(args.device)
     rank = paddle.distributed.get_rank()
